requirements.txt
streamlit
pandas
scikit-learn
numpy
nltk
PyMuPDF
python-docx
pickle-mixin
